+++ jkrimporter/providers/db/dbprovider.py
@@ -56,7 +56,6 @@ from .services.dvv_poimintapvm import (
 )
 from .services.kohde import (
     add_ulkoinen_asiakastieto_for_kohde,
-    create_new_kohde,
     create_perusmaksurekisteri_kohteet,
     find_kohde_by_address,
     find_kohde_by_prt,
@@ -150,7 +149,7 @@ def insert_kuljetukset(
             session.add(db_kuljetus)
 
 
-def find_and_update_kohde(session, asiakas, do_create, do_update_kohde, prt_counts, kitu_counts, address_counts):
+def find_and_update_kohde(session, asiakas, do_update_kohde, prt_counts, kitu_counts, address_counts):
     """
     Etsii olemassa olevan kohteen asiakkaalle tai luo uuden.
     """
@@ -180,24 +179,13 @@ def find_and_update_kohde(session, asiakas, do_create, do_update_kohde, prt_coun
             kohde = find_kohde_by_address(session, asiakas)
             
         if kohde and do_update_kohde:
+            print("Kohde found, updating dates...")
             update_kohde(kohde, asiakas)
-        elif do_create:
-            print("Kohde not found, creating new one...")
-            kohde = create_new_kohde(session, asiakas)
             
         if kohde:
             add_ulkoinen_asiakastieto_for_kohde(session, kohde, asiakas)
         else:
             print("Could not find kohde.")
-
-    # 4. Jos kohde luotu ilman rakennuksia, etsi sopivat rakennukset
-    if do_create and kohde and not kohde.rakennus_collection:
-        print("New kohde created. Looking for buildings...")
-        buildings = find_buildings_for_kohde(
-            session, asiakas, prt_counts, kitu_counts, address_counts
-        )
-        if buildings:
-            kohde.rakennus_collection = buildings
 
     return kohde
 
@@ -220,7 +208,6 @@ def import_asiakastiedot(
     alkupvm: Optional[datetime.date],
     loppupvm: Optional[datetime.date],
     urakoitsija: Tiedontuottaja,
-    do_create: bool,
     do_update_contact: bool,
     do_update_kohde: bool,
     prt_counts: Dict[str, IntervalCounter],
@@ -231,7 +218,6 @@ def import_asiakastiedot(
     kohde = find_and_update_kohde(
         session,
         asiakas,
-        do_create,
         do_update_kohde,
         prt_counts,
         kitu_counts,
@@ -420,7 +406,6 @@ class DbProvider:
         self,
         jkr_data: JkrData,
         tiedontuottaja_lyhenne: str,
-        ala_luo: bool,
         ala_paivita_yhteystietoja: bool,
         ala_paivita_kohdetta: bool,
         siirtotiedosto: Path,
@@ -479,7 +464,6 @@ class DbProvider:
                         jkr_data.alkupvm,
                         jkr_data.loppupvm,
                         tiedontuottajat[urakoitsija_tunnus],
-                        not ala_luo,
                         not ala_paivita_yhteystietoja,
                         not ala_paivita_kohdetta,
                         prt_counts,
@@ -758,7 +742,7 @@ class DbProvider:
                             for prt in kohdentumattomat_prt:
                                 for rawdata in ilmoitus.rawdata:
                                     if rawdata.get(
-                                        "Rakennuksen tiedot, jossa kompostori sijaitsee:Käsittelijän lisäämä tunniste"
+                                        "1. Kompostoria käyttävän rakennuksen tiedot:Käsittelijän lisäämä tunniste"
                                     ) == prt:
                                         kohdentumattomat.append(rawdata)
                     else: